Robust uppdates
2024 robust updates towards final release

+++ lib/auth_screens/sign_up_verify_email.dart
@@ -2,7 +2,10 @@ import 'package:bars/utilities/exports.dart';
 
 class SignpsScreenVerifyEmail extends StatefulWidget {
   static final id = 'Signup_screen';
-
+  final String email;
+  SignpsScreenVerifyEmail({
+    required this.email,
+  });
   @override
   _SignpsScreenVerifyEmailState createState() =>
       _SignpsScreenVerifyEmailState();
@@ -18,7 +21,6 @@ class _SignpsScreenVerifyEmailState extends State<SignpsScreenVerifyEmail>
 
   final _formKey = GlobalKey<FormState>();
   bool _isLoading = false;
-  bool _isVerified = false;
   Timer? timer;
 
   @override
@@ -49,17 +51,32 @@ class _SignpsScreenVerifyEmailState extends State<SignpsScreenVerifyEmail>
 
   _isVerifiedUSer() async {
     final _auth = FirebaseAuth.instance;
-
-    _isVerified = FirebaseAuth.instance.currentUser!.emailVerified;
-
-    if (!_isVerified) {
-      _submit();
-      timer = Timer.periodic(Duration(seconds: 3), (_) => checkEmailVerified());
+    final user = _auth.currentUser;
+    Future.delayed(Duration(milliseconds: 500));
+    try {
+      if (user != null && !user.emailVerified) {
+        await user.sendEmailVerification();
+        print('email sent');
+        mySnackBar(
+          context,
+          "Verification sent to:\n${widget.email.toLowerCase()}",
+        );
+      }
+    } catch (e) {
+      String error = e.toString();
+      String result = error.contains(']')
+          ? error.substring(error.lastIndexOf(']') + 1)
+          : error;
+      mySnackBar(context, 'Request Failed\n$result,');
     }
 
-    Timer(Duration(minutes: 4), () async {
-      if (mounted) {
-        if (!_isVerified) {
+    if (user != null && !user.emailVerified) {
+      timer = Timer.periodic(Duration(seconds: 5), (_) => checkEmailVerified());
+      Timer(Duration(minutes: 4), () async {
+        if (mounted && !user.emailVerified) {
+          Provider.of<UserData>(context, listen: false).setName(
+            '',
+          );
           await _auth.currentUser!.delete().then((value) async {
             await _auth.signOut();
             Navigator.pushAndRemoveUntil(
@@ -69,274 +86,132 @@ class _SignpsScreenVerifyEmailState extends State<SignpsScreenVerifyEmail>
                 ),
                 (Route<dynamic> route) => false);
           });
+
+          mySnackBar(
+            context,
+            "Registration cancelled\nFailed to verify email.",
+          );
         }
-      }
-      Flushbar(
-        maxWidth: MediaQuery.of(context).size.width,
-        backgroundColor: Colors.white,
-        margin: EdgeInsets.all(8),
-        flushbarPosition: FlushbarPosition.TOP,
-        flushbarStyle: FlushbarStyle.FLOATING,
-        titleText: Text(
-          'Registration cancelled.',
-          style: TextStyle(
-            color: Colors.black,
-          ),
-        ),
-        icon: Icon(
-          MdiIcons.checkCircleOutline,
-          size: 30.0,
-          color: Colors.blue,
-        ),
-        messageText: RichText(
-          textScaleFactor: MediaQuery.of(context).textScaleFactor,
-          text: TextSpan(children: [
-            TextSpan(
-                text: "Failed to verify email.",
-                style: TextStyle(fontSize: 14, color: Colors.black)),
-          ]),
-        ),
-        duration: Duration(seconds: 2),
-        isDismissible: true,
-        leftBarIndicatorColor: Colors.blue,
-      )..show(context);
+      });
+    }
+  }
+
+  Future checkEmailVerified() async {
+    final _auth = FirebaseAuth.instance;
+    final user = _auth.currentUser;
+    await user!.reload();
+    print('checking email');
+
+    if (user.emailVerified) {
+      timer?.cancel();
+      setState(() {
+        _isLoading = true;
+      });
+
+      return AuthService.verifyUseer(
+        context,
+        FirebaseAuth.instance.currentUser!.email!,
+      );
+    }
+    setState(() {
+      _isLoading = false;
     });
   }
+
+//
 
   _submit() async {
     final user = FirebaseAuth.instance.currentUser;
     await user!.sendEmailVerification();
-    Flushbar(
-      maxWidth: MediaQuery.of(context).size.width,
-      backgroundColor: Colors.white,
-      margin: EdgeInsets.all(8),
-      flushbarPosition: FlushbarPosition.TOP,
-      flushbarStyle: FlushbarStyle.FLOATING,
-      titleText: Text(
-        'Done.',
-        style: TextStyle(
-          color: Colors.black,
-        ),
-      ),
-      icon: Icon(
-        MdiIcons.checkCircleOutline,
-        size: 30.0,
-        color: Colors.blue,
-      ),
-      messageText: RichText(
-        textScaleFactor: MediaQuery.of(context).textScaleFactor,
-        text: TextSpan(children: [
-          TextSpan(
-              text: "Verification sent.",
-              style: TextStyle(fontSize: 14, color: Colors.black)),
-        ]),
-      ),
-      duration: Duration(seconds: 2),
-      isDismissible: true,
-      leftBarIndicatorColor: Colors.blue,
-    )..show(context);
-  }
-
-  Future checkEmailVerified() async {
-    await FirebaseAuth.instance.currentUser!.reload();
-if(mounted){
- setState(() {
-      _isVerified = FirebaseAuth.instance.currentUser!.emailVerified;
-    });
-
-}
-   
-    if (_isVerified) {
-      timer?.cancel();
-      return AuthService.verifyUseer(
-        context,
-        Provider.of<UserData>(context, listen: false).post3,
-        Provider.of<UserData>(context, listen: false).post1,
-      );
-    }
+    mySnackBar(
+      context,
+      "Verification sent to:\n${widget.email.toLowerCase()}",
+    );
   }
 
   @override
   void dispose() {
-    timer!.cancel();
+    if (timer != null) {
+      timer!.cancel();
+    }
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    String email = Provider.of<UserData>(context, listen: false).post1;
-    final double width = Responsive.isDesktop(context)
-        ? 600.0
-        : MediaQuery.of(context).size.width;
+    final double width = MediaQuery.of(context).size.width;
 
     return AnimatedBuilder(
         animation: animationController,
         builder: (BuildContext context, Widget? child) {
           return Scaffold(
             backgroundColor: Color(0xFF1a1a1a),
-            body: Align(
-              alignment: Alignment.center,
-              child: Container(
-                width: width,
-                child: GestureDetector(
-                  onTap: () => FocusScope.of(context).unfocus(),
-                  child: SingleChildScrollView(
-                    child: SafeArea(
-                      child: Container(
-                        height: MediaQuery.of(context).size.height,
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.center,
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          children: <Widget>[
-                            FadeAnimation(
-                              1,
-                              Icon(
-                                Icons.mail_outline,
-                                size: 100,
+            body: GestureDetector(
+              onTap: () => FocusScope.of(context).unfocus(),
+              child: SingleChildScrollView(
+                child: SafeArea(
+                  child: Container(
+                    width: width,
+                    height: MediaQuery.of(context).size.height,
+                    child: Form(
+                      key: _formKey,
+                      child: Column(
+                        crossAxisAlignment: CrossAxisAlignment.center,
+                        mainAxisAlignment: MainAxisAlignment.center,
+                        children: <Widget>[
+                          Icon(
+                            Icons.mail_outline,
+                            size:
+                                ResponsiveHelper.responsiveHeight(context, 100),
+                            color: Colors.white,
+                          ),
+                          Padding(
+                            padding: const EdgeInsets.all(30.0),
+                            child: Text(
+                              'A verification link has been sent to ${widget.email.toLowerCase()}. Sign in to your email and verify your account to register',
+                              style: TextStyle(
                                 color: Colors.white,
+                                fontSize: ResponsiveHelper.responsiveFontSize(
+                                    context, 14.0),
+                              ),
+                              textAlign: TextAlign.center,
+                            ),
+                          ),
+                          const SizedBox(height: 40.0),
+                          AnimatedContainer(
+                            duration: Duration(milliseconds: 300),
+                            height: _isLoading ? 0.0 : 250,
+                            width: double.infinity,
+                            curve: Curves.easeInOut,
+                            child: SingleChildScrollView(
+                              child: Column(
+                                children: [
+                                  Hero(
+                                    tag: 'Sign Up',
+                                    child: AlwaysWhiteButton(
+                                      textColor: Colors.black,
+                                      buttonText: 'Re-send ',
+                                      onPressed: () {
+                                        _submit();
+                                      },
+                                      buttonColor: Colors.white,
+                                    ),
+                                  ),
+                                ],
                               ),
                             ),
-                            Form(
-                              key: _formKey,
-                              child: AutofillGroup(
-                                child: Column(
-                                  mainAxisSize: MainAxisSize.min,
-                                  children: <Widget>[
-                                    Padding(
-                                      padding: const EdgeInsets.all(30.0),
-                                      child: Text(
-                                        'A verification link has been sent to $email. Sign in to your email and verify your account to register',
-                                        style: TextStyle(
-                                          color: Colors.white,
-                                        ),
-                                        textAlign: TextAlign.center,
-                                      ),
-                                    ),
-                                    SizedBox(height: 40.0),
-                                    AnimatedContainer(
-                                      duration: Duration(milliseconds: 300),
-                                      height: _isLoading ? 0.0 : 250,
-                                      width: double.infinity,
-                                      curve: Curves.easeInOut,
-                                      child: SingleChildScrollView(
-                                        child: Column(
-                                          children: [
-                                            Hero(
-                                              tag: 'Sign Up',
-                                              child: Container(
-                                                width: 250.0,
-                                                child: ElevatedButton(
-                                                  style:
-                                                      ElevatedButton.styleFrom(
-                                                    backgroundColor: Colors.white,
-                                                    elevation: 20.0,
-                                                    foregroundColor: Colors.blue,
-                                                    shape:
-                                                        RoundedRectangleBorder(
-                                                      borderRadius:
-                                                          BorderRadius.circular(
-                                                              20.0),
-                                                    ),
-                                                  ),
-                                                  onPressed: _submit,
-                                                  child: Padding(
-                                                    padding:
-                                                        const EdgeInsets.all(
-                                                            8.0),
-                                                    child: Text(
-                                                      'Re-send ',
-                                                      style: TextStyle(
-                                                        color: Colors.black,
-                                                        fontSize: width > 800
-                                                            ? 24
-                                                            : 16,
-                                                      ),
-                                                    ),
-                                                  ),
-                                                ),
-                                              ),
-                                            ),
-                                            // SizedBox(
-                                            //     height:
-                                            //         width > 800 ? 40.0 : 20),
-                                            // FadeAnimation(
-                                            //   2,
-                                            //   Container(
-                                            //     width: 250.0,
-                                            //     child: OutlinedButton(
-                                            //       style:
-                                            //           OutlinedButton.styleFrom(
-                                            //         primary: Colors.blue,
-                                            //         side: BorderSide(
-                                            //           width: 1.0,
-                                            //           color: Colors.white,
-                                            //         ),
-                                            //         shape:
-                                            //             RoundedRectangleBorder(
-                                            //           borderRadius:
-                                            //               BorderRadius.circular(
-                                            //                   20.0),
-                                            //         ),
-                                            //       ),
-                                            //       child: Padding(
-                                            //         padding:
-                                            //             const EdgeInsets.all(
-                                            //                 8.0),
-                                            //         child: Text(
-                                            //           'Back',
-                                            //           style: TextStyle(
-                                            //             color: Colors.white,
-                                            //             fontSize: width > 800
-                                            //                 ? 24
-                                            //                 : 16,
-                                            //           ),
-                                            //         ),
-                                            //       ),
-                                            //       onPressed: () =>
-                                            //           Navigator.pop(context),
-                                            //     ),
-                                            //   ),
-                                            // ),
-                                            // SizedBox(height: 80.0),
-                                            // FadeAnimation(
-                                            //   2,
-                                            //   GestureDetector(
-                                            //     onTap: () =>
-                                            //         Navigator.pushNamed(
-                                            //             context, Password.id),
-                                            //     child: Text('Forgot Password?',
-                                            //         style: TextStyle(
-                                            //           color: Colors.blueGrey,
-                                            //           fontSize:
-                                            //               width > 800 ? 18 : 12,
-                                            //         ),
-                                            //         textAlign: TextAlign.right),
-                                            //   ),
-                                            // ),
-                                          ],
-                                        ),
-                                      ),
-                                    ),
-                                    Provider.of<UserData>(context,
-                                                listen: false)
-                                            .isLoading
-                                        ? SizedBox(
-                                            height: 0.5,
-                                            child: LinearProgressIndicator(
-                                              backgroundColor:
-                                                  Colors.transparent,
-                                              valueColor:
-                                                  AlwaysStoppedAnimation(
-                                                      Colors.white),
-                                            ),
-                                          )
-                                        : const SizedBox.shrink()
-                                  ],
-                                ),
-                              ),
-                            )
-                          ],
-                        ),
+                          ),
+                          _isLoading
+                              ? SizedBox(
+                                  height: 0.5,
+                                  child: LinearProgressIndicator(
+                                    backgroundColor: Colors.transparent,
+                                    valueColor:
+                                        AlwaysStoppedAnimation(Colors.white),
+                                  ),
+                                )
+                              : const SizedBox.shrink()
+                        ],
                       ),
                     ),
                   ),
